refactor(onboarding): merge auth method step into SAM deploy
Removes the standalone CippAuthMethodStep wizard step and moves the authentication method radio selector directly into the CippSAMDeploy component. Also adds certificateAuth and certificateOnly fields to the OpenAPI spec, and minor UI fixes in the SharePoint library copy dialog.
