+++ frontend/src/components/CippWizard/CippSAMDeploy.jsx
@@ -1,7 +1,9 @@
 import { useEffect, useState } from "react";
-import { Alert, Stack, Box, Link } from "@mui/material";
+import { Alert, Stack, Box, Link, Typography } from "@mui/material";
+import { useWatch } from "react-hook-form";
 import { CIPPM365OAuthButton } from "../CippComponents/CIPPM365OAuthButton";
 import { CippApiResults } from "../CippComponents/CippApiResults";
+import CippFormComponent from "../CippComponents/CippFormComponent";
 import { ApiPostCall } from "../../api/ApiCall";
 import { CippWizardStepButtons } from "./CippWizardStepButtons";
 
@@ -12,6 +14,8 @@ export const CippSAMDeploy = (props) => {
     error: null,
     loading: false,
   });
+
+  const authMethod = useWatch({ control: formControl.control, name: "authMethod" });
 
   // Block next step until SAM app is created
   formControl.register("SAMWizard", {
@@ -85,6 +89,33 @@ export const CippSAMDeploy = (props) => {
 
   return (
     <Stack spacing={2}>
+      <Stack spacing={2}>
+        <Typography variant="h6" id="auth-method-heading">
+          Authentication method
+        </Typography>
+        <Typography variant="body2" color="text.secondary">
+          Choose how CIPP authenticates its application registration to Microsoft. This determines
+          whether a client secret is created during setup.
+        </Typography>
+        <CippFormComponent
+          type="radio"
+          name="authMethod"
+          formControl={formControl}
+          defaultValue="certificate"
+          aria-labelledby="auth-method-heading"
+          options={[
+            {
+              value: "certificate",
+              label:
+                "Certificate (recommended) - no client secret; CIPP generates and auto-rotates it.",
+            },
+            {
+              value: "secret",
+              label: "Client secret - a shared secret; newer Entra tenants may block creating one.",
+            },
+          ]}
+        />
+      </Stack>
       <Alert severity="info">
         To run this setup you will need the following prerequisites:
         <li>
